image path moved to config folder
