+++ lib/ui/on_board_screen/on_board_data_model.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../image_path.dart';
+import '../../config/image_path.dart';
 
 class OnBoardDataModel {
   final String onBoardTitle;
